fix: use wagmi transports setup in viem client
