+++ lib/modules/web3/WagmiConfig.tsx
@@ -1,11 +1,9 @@
 'use client'
 
-import { Chain, connectorsForWallets } from '@rainbow-me/rainbowkit'
+import { connectorsForWallets } from '@rainbow-me/rainbowkit'
 
-import { createConfig, fallback, http } from 'wagmi'
+import { createConfig } from 'wagmi'
 
-import { getGqlChain } from '@/lib/config/app.config'
-import { SupportedChainId } from '@/lib/config/config.types'
 import { getProjectConfig } from '@/lib/config/getProjectConfig'
 import {
   coinbaseWallet,
@@ -16,18 +14,8 @@ import {
   safeWallet,
   walletConnectWallet,
 } from '@rainbow-me/rainbowkit/wallets'
-import { chains, rpcOverrides } from './ChainConfig'
-
-function getTransports(chain: Chain) {
-  const gqlChain = getGqlChain(chain.id as SupportedChainId)
-  const overrideRpcUrl = rpcOverrides[gqlChain]
-  const fallbackRpcUrl = rpcOverrides[gqlChain]
-  return fallback([http(overrideRpcUrl), http(), http(fallbackRpcUrl)])
-}
-
-const transports = Object.fromEntries(
-  chains.map(chain => [chain.id, getTransports(chain)])
-) as Record<number, ReturnType<typeof getTransports>>
+import { chains } from './ChainConfig'
+import { transports } from './transports'
 
 const appName = getProjectConfig().projectName
 const projectId = process.env.NEXT_PUBLIC_WALLET_CONNECT_ID || ''
