+++ lib/shared/services/viem/viem.client.ts
@@ -1,8 +1,9 @@
-import { createPublicClient, http } from 'viem'
+import { createPublicClient } from 'viem'
 import { GqlChain } from '../api/generated/graphql'
 import { getNetworkConfig } from '@/lib/config/app.config'
-import { chains } from '@/lib/modules/web3/ChainConfig'
+import { chains, chainsByKey } from '@/lib/modules/web3/ChainConfig'
 import { Chain } from 'viem'
+import { getTransports } from '@/lib/modules/web3/transports'
 
 function getViemChain(chainId: number): Chain {
   const chain = chains.find(chain => chain.id === chainId)
@@ -15,6 +16,6 @@ export function getViemClient(chain: GqlChain) {
 
   return createPublicClient({
     chain: getViemChain(chainId),
-    transport: http(),
+    transport: getTransports(chainsByKey[chainId]),
   })
 }
